// src/renderer/src/types/telemetry.ts
export interface Telemetry {
  timestamp: string;
  values: Values;
}